Restructure components. Scaffold tests
