+++ packages/lib-classifier/src/components/Classifier/components/FieldGuide/FieldGuideContainer.spec.js
@@ -1,14 +1,7 @@
 import { shallow, mount } from 'enzyme'
 import React from 'react'
-import FieldGuideButton from './FieldGuideButton'
+import FieldGuideContainer from './FieldGuideContainer'
 
-describe('Component > FieldGuideButton', function () {
-  it('should render without crashing', function () {
-    shallow(<FieldGuideButton />)
-  })
+describe('Component > FieldGuideContainer', function () {
 
-  it('should have an ARIA label', function () {
-    const wrapper = mount(<FieldGuideButton />)
-    expect(wrapper.find('button').prop('aria-label')).to.equal('Open Field Guide')
-  })
 })
