+++ packages/lib-classifier/src/components/Classifier/components/FieldGuide/components/FieldGuide/components/FieldGuideItem/index.js
@@ -1,0 +1,1 @@
+export { default } from './FieldGuideItem'
