+++ packages/lib-classifier/src/components/Classifier/components/FieldGuide/components/FieldGuide/components/FieldGuideItems/index.js
@@ -1,0 +1,1 @@
+export { default } from './FieldGuideItems'
