+++ packages/lib-classifier/src/components/Classifier/components/FieldGuide/components/FieldGuide/components/FieldGuideItems/locales/en.json
@@ -1,5 +1,5 @@
 {
-  "FieldGuideItem": {
+  "FieldGuideItemButton": {
     "ariaTitle": "Select %(title)s"
   }
 }
